error handing, tested vamana in main and changed some variable names

+++ src/b2vecs_read.cpp
@@ -59,7 +59,7 @@ vector<vector<float>> b2fvecs_read(const char* filename, size_t a, size_t b)
     // Reshape the buffer into a 2D vector, ignoring the first 4 floats (dimension) of each vector
     vector<vector<float>> vectors(n, vector<float>(d));
     for (size_t i = 0; i < n; ++i) {
-        for (size_t j = 0; j < d; ++j) {
+        for (size_t j = 0; j < static_cast<size_t>(d); ++j) {
             vectors[i][j] = buffer[i * (d + 4) + j + 4];  // Skip the first 4 floats (dimension)
         }
     }
